online-error-handling-tool/requirements.txt
# Core Web Framework
Flask==3.1.2
Flask-CORS==4.0.0
Werkzeug==3.1.3
Jinja2==3.1.6
MarkupSafe==3.0.3
click==8.3.0
itsdangerous==2.2.0
blinker==1.9.0

# Image Processing
numpy==2.2.6
Pillow==11.0.0
opencv-python==4.12.0.88
tifffile==2025.5.10
scikit-image==0.24.0
scipy==1.14.1
